Refactor get_planes_data function
Point to the FAA data source that is a single zip file and have 'year' be NULL but retain for compatibility across other functions. Adjust the process_planes_master function to not delete the temporary folder holding this file so it can be accessed by the process_planes_ref function.